Fix tests
There are a couple of things here.
* firstly, blinker wasn't in your requirements file. I've updated that
for you
* second, all your fixtures were scoped to the 'function'. That meant
that your database was destroyed after every test. Unfortunately it
seems the order was destroy database --> try to add things to the
database --> create database. I've fixed this by scoping the db,
application, and client to the entire testing session. You've already
made a good start on this pattern by tearing down data you create in
your fixtures
* third, your tests are still failing. I think they're failing for good
reasons though