Add start page and create draft masterclass
This commit also includes a test which checks that a draft masterclass is created upon posting to create-masterclass, and that its id is added to the session. As part of this I have separated the test_app and the test_client, created a logged_in_user fixture and created a blank session fixture which will allow for this and future tests to interact with the db session.