Choose ddat job family and add choice to session
The chosen job family is added to the session so that when a user goes back to the next step in the form, in which they're displayed the content items related to that job family, we have access to what option they originally chose. The job family radio buttons are extracted into their own template as they will be reused when a user adds new masterclass content.