Display existing content and add to draft on post
Existing content for the job family in the session is displayed. If a user chooses an existing one, the id for that item is added to the draft masterclass, if not a user is directed to the 'create new' page. I have included a test which checks if the content id is successfully added to the draft masterclass.